MPI_Comm abstraction and tests.
- Removed dependency on mpi.h for all files but
  abstraction implementation.
- Created CommFactory singleton that is exposed via
  C++ interface.
- Added new exception type and message.
- Added MPI implementation unit test suite.

+++ src/Profile.cpp
@@ -61,6 +61,7 @@
 #include "ProfileTable.hpp"
 #include "geopm_sched.h"
 #include "config.h"
+#include "Comm.hpp"
 
 static bool geopm_prof_compare(const std::pair<uint64_t, struct geopm_prof_message_s> &aa, const std::pair<uint64_t, struct geopm_prof_message_s> &bb)
 {
@@ -69,7 +70,9 @@ static bool geopm_prof_compare(const std::pair<uint64_t, struct geopm_prof_messa
 
 static geopm::Profile &geopm_default_prof(void)
 {
-    static geopm::Profile default_prof(geopm_env_profile(), MPI_COMM_WORLD);
+    geopm::IComm *tmp_comm = geopm::geopm_get_comm(geopm::MPICOMM_DESCRIPTION);
+    static geopm::Profile default_prof(geopm_env_profile(), tmp_comm);
+    delete tmp_comm;
     return default_prof;
 }
 
@@ -210,7 +213,7 @@ extern "C"
 namespace geopm
 {
 
-    Profile::Profile(const std::string prof_name, MPI_Comm comm)
+    Profile::Profile(const std::string prof_name, IComm *comm)
         : m_is_enabled(true)
         , m_prof_name(prof_name)
         , m_curr_region_id(0)
@@ -226,7 +229,7 @@ namespace geopm
         , m_tprof_table(NULL)
         , M_OVERHEAD_FRAC(0.01)
         , m_scheduler(NULL)
-        , m_shm_comm(MPI_COMM_NULL)
+        , m_shm_comm(NULL)
         , m_rank(0)
         , m_shm_rank(0)
         , m_is_first_sync(true)
@@ -244,11 +247,11 @@ namespace geopm
         int shm_num_rank = 0;
 
         m_scheduler = new SampleScheduler(M_OVERHEAD_FRAC);
-        MPI_Comm_rank(comm, &m_rank);
-        geopm_comm_split_shared(comm, "prof", &m_shm_comm);
-        PMPI_Comm_rank(m_shm_comm, &m_shm_rank);
-        PMPI_Comm_size(m_shm_comm, &shm_num_rank);
-        PMPI_Barrier(m_shm_comm);
+        m_rank = comm->rank();
+        m_shm_comm = geopm_get_comm(comm, "prof", IComm::M_COMM_SPLIT_TYPE_SHARED);
+        m_shm_rank = m_shm_comm->rank();
+        shm_num_rank = m_shm_comm->num_rank();
+        m_shm_comm->barrier();
 
         std::string key(geopm_env_shmkey());
         key += "-sample";
@@ -265,7 +268,7 @@ namespace geopm
             m_is_enabled = false;
             return;
         }
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         if (!m_shm_rank) {
             m_ctl_shmem->unlink();
         }
@@ -299,7 +302,7 @@ namespace geopm
                     }
                 }
             }
-            PMPI_Barrier(m_shm_comm);
+            m_shm_comm->barrier();
         }
 
         if (!m_shm_rank) {
@@ -324,8 +327,11 @@ namespace geopm
         std::ostringstream table_shm_key;
         table_shm_key << key <<  "-"  << m_rank;
         m_table_shmem = new SharedMemoryUser(table_shm_key.str(), 3.0);
-        PMPI_Barrier(m_shm_comm);
-        m_table_shmem->unlink();
+        m_shm_comm->barrier();
+        if (!m_shm_rank) {
+            m_table_shmem->unlink();
+        }
+        m_shm_comm->barrier();
 
         m_table_buffer = m_table_shmem->pointer();
         m_table = new ProfileTable(m_table_shmem->size(), m_table_buffer);
@@ -333,7 +339,7 @@ namespace geopm
         std::string tprof_key(geopm_env_shmkey());
         tprof_key += "-tprof";
         m_tprof_shmem = new SharedMemoryUser(tprof_key, 3.0);
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         if (!m_shm_rank) {
             m_tprof_shmem->unlink();
         }
@@ -430,7 +436,7 @@ namespace geopm
         if (!m_curr_region_id && region_id) {
             if (!geopm_region_id_is_mpi(region_id) &&
                 geopm_env_do_region_barrier()) {
-                PMPI_Barrier(m_shm_comm);
+                m_shm_comm->barrier();
             }
             m_curr_region_id = region_id;
             m_num_enter = 0;
@@ -493,7 +499,7 @@ namespace geopm
         if (!m_num_enter) {
             if (!geopm_region_id_is_mpi(region_id) &&
                 geopm_env_do_region_barrier()) {
-                PMPI_Barrier(m_shm_comm);
+                m_shm_comm->barrier();
             }
             if (geopm_region_id_is_mpi(region_id)) {
                 m_curr_region_id = geopm_region_id_set_mpi(m_parent_region);
@@ -559,7 +565,7 @@ namespace geopm
 #endif
 
         struct geopm_prof_message_s sample;
-        PMPI_Barrier(m_shm_comm);
+        m_shm_comm->barrier();
         if (!m_shm_rank) {
             sample.rank = m_rank;
             sample.region_id = GEOPM_REGION_ID_EPOCH;
@@ -657,8 +663,8 @@ namespace geopm
                                      m_overhead_time,
                                      m_overhead_time_shutdown};
         double max_overhead[3] = {};
-        MPI_Reduce(overhead_buffer, max_overhead, 3,
-                   MPI_DOUBLE, MPI_MAX, 0, MPI_COMM_WORLD);
+        IComm *tmp_comm = geopm_get_comm(geopm::MPICOMM_DESCRIPTION);
+        tmp_comm->reduce_max(overhead_buffer, max_overhead, 3, 0);
         if (!m_rank) {
             std::cout << "GEOPM startup (seconds):  " << max_overhead[0] << std::endl;
             std::cout << "GEOPM runtime (seconds):  " << max_overhead[1] << std::endl;
@@ -774,7 +780,7 @@ namespace geopm
         return result;
     }
 
-    void ProfileSampler::sample(std::vector<std::pair<uint64_t, struct geopm_prof_message_s> > &content, size_t &length, MPI_Comm comm)
+    void ProfileSampler::sample(std::vector<std::pair<uint64_t, struct geopm_prof_message_s> > &content, size_t &length, IComm *comm)
     {
         length = 0;
         if (m_ctl_msg->is_sample_begin() ||
